+++ CMakeLists.txt
@@ -155,14 +155,13 @@ project(libtropic_SDK
 set(SDK_SRCS ${SDK_SRCS}
     ${CMAKE_CURRENT_SOURCE_DIR}/src/libtropic.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_crc16.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l1_port_wrap.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_port_wrap.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l1.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/libtropic_l2.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l2_frame_check.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l3_process.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/libtropic_l3.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_hkdf.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_random.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_asn1_der.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/libtropic_default_sh0_keys.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_tr01_attrs.c
@@ -176,12 +175,11 @@ set(SDK_INCS ${SDK_INCS}
     ${CMAKE_CURRENT_SOURCE_DIR}/include/libtropic_l2.h
     ${CMAKE_CURRENT_SOURCE_DIR}/include/libtropic_l3.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_crc16.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l1_port_wrap.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_port_wrap.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l1.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l2_frame_check.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_l3_process.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_hkdf.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_random.h
     ${CMAKE_CURRENT_SOURCE_DIR}/src/lt_asn1_der.h
 )
 set(SDK_DIRS_PRIV ${SDK_DIRS_PRIV}
