+++ tests/functional/lt_test_ire_write_i_config.c
@@ -12,7 +12,7 @@
 #include "libtropic_common.h"
 #include "libtropic_functional_tests.h"
 #include "libtropic_logging.h"
-#include "lt_random.h"
+#include "lt_port_wrap.h"
 #include "string.h"
 
 void lt_test_ire_write_i_config(lt_handle_t *h)
