+++ tests/functional/lt_test_rev_eddsa_sign.c
@@ -14,7 +14,7 @@
 #include "libtropic_functional_tests.h"
 #include "libtropic_logging.h"
 #include "lt_l3_api_structs.h"
-#include "lt_random.h"
+#include "lt_port_wrap.h"
 #include "string.h"
 
 // Pre-generated key for testing using OpenSSL
