+++ tests/functional/lt_test_rev_random_value_get.c
@@ -12,7 +12,7 @@
 #include "libtropic_common.h"
 #include "libtropic_functional_tests.h"
 #include "libtropic_logging.h"
-#include "lt_random.h"
+#include "lt_port_wrap.h"
 #include "string.h"
 
 #define RANDOM_VALUE_GET_LOOPS 300
