+++ tests/functional/lt_test_rev_startup_req.c
@@ -13,8 +13,8 @@
 #include "libtropic_functional_tests.h"
 #include "libtropic_l2.h"
 #include "libtropic_logging.h"
-#include "lt_l1_port_wrap.h"
 #include "lt_l2_api_structs.h"
+#include "lt_port_wrap.h"
 
 #define REBOOT_WAIT_ATTEMPTS 10
